fix Перенес функционал логаута, так как ломал переход к регистрации
