+++ packages/client/src/hooks/useAuthGuard.ts
@@ -1,36 +1,16 @@
-import { useCallback, useEffect } from 'react';
+import { useEffect } from 'react';
 import { useLocation, useNavigate } from 'react-router-dom';
 import { SigninPagePath } from '@/router/paths';
-import { getUserInfo, oauthSignIn } from '@/store/user/user.actions';
+import { oauthSignIn } from '@/store/user/user.actions';
 import { REDIRECT_URI } from '@/сonstants/main';
-import { useAppDispatch, useAppSelector } from '@/hooks/index';
-import { showUserData } from '@/store/user/user.slice';
+import { useAppDispatch } from '@/hooks/index';
 
 export const useAuthGuard = () => {
   const navigate = useNavigate();
   const { pathname } = useLocation();
   const dispatch = useAppDispatch();
 
-  const { profile: user } = useAppSelector(showUserData);
   const isLoggedIn = Boolean(localStorage.getItem('user_in'));
-
-  const fetchUserData = useCallback(async () => {
-    const resultAction = await dispatch(getUserInfo());
-    if (getUserInfo.fulfilled.match(resultAction)) {
-      const { payload } = resultAction;
-      return payload;
-    }
-    return null;
-  }, [dispatch]);
-
-  useEffect(() => {
-    fetchUserData().then((payload) => {
-      if (!payload) {
-        localStorage.clear();
-        navigate('/auth/login', { replace: true });
-      }
-    });
-  }, [user.id, navigate, fetchUserData]);
 
   useEffect(() => {
     if (!pathname.includes('/auth')) {
@@ -40,7 +20,6 @@ export const useAuthGuard = () => {
       if (window.location.search.includes('code')) {
         const yaCode: string = window.location.search.split('=')[1];
 
-        // @ts-ignore
         dispatch(oauthSignIn({ code: yaCode, redirect_uri: REDIRECT_URI }));
         navigate(SigninPagePath.path);
       }
